server/requirements.txt
uvicorn
fastapi
pydantic
resend
pydantic[email]
python-dotenv
email-validator
requests